Allow user to move caret to end of a line or its beginning
On Windows the user can press the 'home' or 'end' key to jump around. On
OSX he has to use cmd+arrow keys to move around.

If the user holds down shift while performing the above operations the
whole line will be selected from where the caret was currently placed.