Launcher for the LuaJIT backend.
Looks for main.lua. Launch it like this:

    luajit myapp/lib/luigi/launch.lua

If luigi isn't inside the project directory, pass
the path containing main.lua as the second argument.
The path must end with a directory separator.

    luajit /opt/luigi/launch.lua ./myapp/

If the app prefixes luigi modules with something
other then 'luigi', pass that prefix as the third
argument.

    luajit /opt/luigi/launch.lua ./myapp/ lib.luigi